requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
Flask-Marshmallow>=1.2.0
marshmallow-sqlalchemy>=1.0.0
Flask-Bcrypt>=1.0.1
python-dotenv>=1.0.0
flask-cors>=4.0.0
